adding rpcs and visualization
